Refactor StackReferenceMock filtering in Mocks.cs
Modified the filtering of StackReferenceMock instances in Mocks.cs in order to simplify and improve code readability. Refactored the two separate Where clauses into OfType<StackReferenceMock>() followed by a Where clause, which provides the same functionality but in a more clear and concise manner. Also removed redundant newlines at the end of the file.